+++ tests/TestCase.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace graychen\yii2\queue\backend\tests;
 
 use Yii;
@@ -50,9 +51,10 @@ class TestCase extends \PHPUnit_Framework_TestCase
                 ],
                 'redis' => [
                     'class' => 'yii\redis\Connection',
-                    'hostname' => '127.0.0.1',
-                    'password' => null,
-                    'database' => 0
+                    'hostname' => 'localhost',
+                    'port' => 6379,
+                    'database' => 0,
+                    'password' => null
                 ],
                 'queue' => [
                     'class' => 'yii\queue\redis\Queue',
@@ -77,6 +79,7 @@ class TestCase extends \PHPUnit_Framework_TestCase
     {
         return dirname(__DIR__) . '/vendor';
     }
+
     /**
      * Destroys application in Yii::$app by setting it to null.
      */